fixed updating
Former-commit-id: 90a4ec7fbd1de2422fc81c4263021fabe32dae9b [formerly 67f3264389308e4cfea190d8cad69d8c0a72cfcb]
Former-commit-id: 1d173d62359c0be15b636fc8dcc93de2bd560234

+++ drugstone/management/commands/populate_db.py
@@ -12,8 +12,6 @@ from drugstone.management.includes.NodeCache import NodeCache
 from drugstone.management.includes import DatasetLoader
 
 
-
-
 class DatabasePopulator:
     def __init__(self, data_dir):
         self.data_dir = data_dir
@@ -93,10 +91,10 @@ def populate(kwargs):
 
     db_populator = DatabasePopulator(data_dir=data_dir)
 
-    if kwargs['clear']:
+    if 'clear' in kwargs and kwargs['clear']:
         db_populator.delete_all()
 
-    if kwargs['delete_model'] is not None:
+    if 'delete_model' in kwargs and kwargs['delete_model'] is not None:
         model_list = kwargs['delete_model'].split(',')
         db_populator.delete_models(model_list)
 
@@ -105,7 +103,8 @@ def populate(kwargs):
     importer = NedrexImporter(nedrex_api_url, cache)
     populator = DataPopulator(cache)
 
-    if kwargs['all']:
+    total_n = 0
+    if 'all' in kwargs and kwargs['all']:
         kwargs['drugs'] = True
         kwargs['disorders'] = True
         kwargs['proteins'] = True
@@ -118,24 +117,29 @@ def populate(kwargs):
     if kwargs['drugs']:
         print('Populating Drugs...')
         n = NedrexImporter.import_drugs(importer, update)
+        total_n +=n
         print(f'Populated {n} Drugs.')
 
     if kwargs['disorders']:
         print('Populating Disorders...')
         n = NedrexImporter.import_disorders(importer, update)
+        total_n += n
         print(f'Populated {n} Disorders.')
 
     if kwargs['proteins']:
         print('Populating Proteins...')
         n = NedrexImporter.import_proteins(importer, update)
+        total_n += n
         print(f'Populated {n} Proteins.')
         print('Populating ENSG IDs...')
         n = DataPopulator.populate_ensg(populator, update)
+        total_n += n
         print(f'Populated {n} ENSG IDs.')
 
     if kwargs['exp']:
         print('Populating Expressions...')
         n = DataPopulator.populate_expressions(populator, update)
+        total_n += n
         print(f'Populated {n} Expressions.')
 
     if kwargs['protein_protein']:
@@ -143,17 +147,21 @@ def populate(kwargs):
         n = NedrexImporter.import_protein_protein_interactions(importer,
                                                                DatasetLoader.get_ppi_nedrex(nedrex_api_url),
                                                                update)
+        total_n += n
         print(f'Imported {n} PPIs from NeDRexDB')
         print('Populating PPIs from STRING...')
         n = DataPopulator.populate_ppi_string(populator, DatasetLoader.get_ppi_string(), update)
+        total_n += n
         print(f'Populated {n} PPIs from STRING.')
 
         print('Populating PPIs from APID...')
         n = DataPopulator.populate_ppi_apid(populator, DatasetLoader.get_ppi_apid(), update)
+        total_n += n
         print(f'Populated {n} PPIs from APID.')
 
         print('Populating PPIs from BioGRID...')
         n = DataPopulator.populate_ppi_biogrid(populator, DatasetLoader.get_ppi_biogrid(), update)
+        total_n += n
         print(f'Populated {n} PPIs from BioGRID.')
 
     if kwargs['protein_drug']:
@@ -161,18 +169,22 @@ def populate(kwargs):
         n = NedrexImporter.import_drug_target_interactions(importer,
                                                            DatasetLoader.get_drug_target_nedrex(nedrex_api_url),
                                                            update)
+        total_n += n
         print(f'Imported {n} PDIs from NeDRexDB')
 
         print('Populating PDIs from Chembl...')
         n = DataPopulator.populate_pdi_chembl(populator, DatasetLoader.get_drug_target_chembl(), update)
+        total_n += n
         print(f'Populated {n} PDIs from Chembl.')
 
         print('Populating PDIs from DGIdb...')
         n = DataPopulator.populate_pdi_dgidb(populator, DatasetLoader.get_drug_target_dgidb(), update)
+        total_n += n
         print(f'Populated {n} PDIs from DGIdb.')
 
         print('Populating PDIs from DrugBank...')
         n = DataPopulator.populate_pdi_drugbank(populator, DatasetLoader.get_drug_target_drugbank(), update)
+        total_n += n
         print(f'Populated {n} PDIs from DrugBank.')
 
     if kwargs['protein_disorder']:
@@ -181,9 +193,11 @@ def populate(kwargs):
                                                                 DatasetLoader.get_protein_disorder_nedrex(
                                                                     nedrex_api_url),
                                                                 update)
+        total_n += n
         print(f'Imported {n} PDis from NeDRexDB')
         print('Populating PDis associations from DisGeNET...')
         n = DataPopulator.populate_pdis_disgenet(populator, DatasetLoader.get_disorder_protein_disgenet(), update)
+        total_n += n
         print(f'Populated {n} PDis associations from DisGeNET.')
 
     if kwargs['drug_disorder']:
@@ -191,7 +205,12 @@ def populate(kwargs):
         n = NedrexImporter.import_drug_disorder_indications(importer,
                                                             DatasetLoader.get_drug_disorder_nedrex(nedrex_api_url),
                                                             update)
+        total_n += n
         print(f'Imported {n} DrDis from NeDRexDB')
         print('Populating DrDi indications from DrugBank...')
         n = DataPopulator.populate_drdis_drugbank(populator, DatasetLoader.get_drug_disorder_drugbank(), update)
+        total_n += n
         print(f'Populated {n} DrDi associations from DrugBank.')
+
+    cache.clear()
+    return total_n
